requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
diff-match-patch==20230430
Django==5.1.1
django-background-tasks==1.2.8
django-celery-beat==2.7.0
django-import-export==4.1.1
django-timezone-field==7.0
gunicorn==23.0.0
kombu==5.4.2
packaging==24.1
prompt_toolkit==3.0.48
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.1.1
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
